Add update bar under tag pills, fix macOS build, macOS-only CI
- Fix vault.rs disk-free calc on macOS (u32*u64 type mismatch that broke the mac build)
- Replace top-bar update button with an update bar under the home tag pills; wired to
  the existing update-available event + apply_update command (download, install, restart)
- Fully localise the update bar and previously-hardcoded home-view strings (en + uz);
  add an i18n:changed hook so imperatively-rendered text re-localises on language switch
- CI now builds macOS only (Windows is built locally); installers are hosted on Cloudflare

+++ src-tauri/src/vault.rs
@@ -90,7 +90,7 @@ pub fn disk_free_bytes(path: &PathBuf) -> Option<u64> {
         let c_path = CString::new(path.to_string_lossy().as_bytes()).ok()?;
         let mut stat: libc::statvfs = unsafe { mem::zeroed() };
         let ret = unsafe { libc::statvfs(c_path.as_ptr(), &mut stat) };
-        if ret == 0 { Some(stat.f_bavail * stat.f_frsize as u64) } else { None }
+        if ret == 0 { Some(stat.f_bavail as u64 * stat.f_frsize as u64) } else { None }
     }
     #[cfg(not(any(target_os = "windows", target_os = "macos")))]
     {
